Important BUG FIX
The bug regards the null move pruning, after which we previously returned a mate if we found one. Since it came from a null move search though, we can't actually return it because it could be false (it probably is).
Also, the prunings in the search are done only if alpha is not a mating score that indicates that for now we will be checkmated (if we are being checkmated, we want to search everything because the solution for escaping could be anywhere in the tree and we want to make sure we find it)